implemened base details for ui change
