--- src/components/chatroom/chatroom.css
@@ -1,65 +1,0 @@
-.chatroomContainer {
-    width: 100%;
-    height: 94%;
-    display: flex;
-    flex-direction: column;
-}
-
-.recipientName {
-    border-bottom: 1px solid black;
-    width: 100%;
-    height: 2.35rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 1.35rem;
-    font-weight: bold;
-}
-
-.recipientNameLink {
-    font-size: 1.35rem;
-    font-weight: bold;
-    color: black;
-    text-decoration: none;
-}
-
-.recipientNameLink:hover {
-    font-size: 1.35rem;
-    font-weight: bold;
-    color: black;
-    text-decoration: underline;
-}
-
-.messagesContainer {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    height: 90%;
-    overflow-y: auto;
-    gap: .5rem;
-    padding-top: .5rem;
-}
-
-.typingContainer {
-    width: 100%;
-    height: 1.5rem;
-    text-align: left;
-    padding-left: 1rem;
-}
-
-.loadMoreBtn {
-    width: 20%;
-    text-align: center;
-    background-color: #47a9ff;
-    border: none;
-    cursor: pointer;
-}
-
-.loadMoreBtn:hover {
-    width: 20%;
-    text-align: center;
-    background-color: #47a9ff;
-    border: 1px solid black;
-    cursor: pointer;
-}
