--- src/components/chatroom/create-message/create-message.css
@@ -1,34 +1,0 @@
-.createMessageContainer {
-    width: 100%;
-    height: 6%;
-}
-
-.newMessageForm {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-}
-
-.messageTextInput {
-    width: 80%;
-    height: 80%;
-    font-family: 'Source Sans Pro', sans-serif;
-    resize: none;
-}
-
-.messageSendButton {
-    width: 15%;
-    height: 80%;
-    background-color: #47a9ff;
-    border: none;
-}
-
-.messageSendButton:hover {
-    width: 15%;
-    height: 80%;
-    cursor: pointer;
-    background-color: #47a9ff;
-    border: 2px solid black;
-}
