--- src/components/chatroom/messages/message/message.css
@@ -1,34 +1,0 @@
-
-.userMessage {
-   width: 97vw;
-   display: flex;
-   flex-direction: row;
-   justify-content: flex-end;
-}
-
-.otherMessage {
-    width: 97vw;
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-start;
-}
-
-.userText {
-    border: 2px solid #047ae0;
-    max-width: 13rem;
-    color: white;
-    background-color: #047ae0;
-    padding: .22rem;
-    overflow-wrap: break-word;
-    border-radius:.75rem;
-}
-
-.otherText {
-    border: 2px solid lightgray;
-    max-width: 13rem;
-    color: black;
-    background-color: lightgray;
-    padding: .22rem;
-    overflow-wrap: break-word;
-    border-radius:.75rem;
-}
