+++ src/components/dashboard/dashboard.css
@@ -1,0 +1,56 @@
+.dashBoardPage {
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+}
+
+.conversationListContainer {
+    border: 1px solid black;
+    width: 30%;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-start;
+    align-items: center;
+    gap: 1rem;
+}
+
+.conversationContainer {
+    width: 70%;
+    height: 100%;
+}
+
+@media only screen and (max-width: 600px) {
+    .conversationContainer {
+        width: 100%;
+    }
+
+    .conversationListContainer {
+        display: none;
+    }
+}
+
+
+.recipientNameContainer {
+    height: 5%;
+}
+
+.messagesContainer {
+    height: 90%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    overflow-y: auto;
+}
+
+.createMessageContainer {
+    height: 5%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
