+++ src/components/dashboard/dashboard.tsx
@@ -1,19 +1,29 @@
 import { useContext, useEffect, useState } from 'react';
 import './dashboard.css';
-import ConversationList from './conversationlist/conversationlist';
+import ConversationList from './list/conversationlist';
 import useFetch from '../../hooks/use-fetch';
 import AuthContext from '../../contexts/authcontext';
 import { ChatInterface } from '../../shared/interfaces/interfaces';
 import { Outlet, useNavigate } from 'react-router';
 
 const Dashboard = () => {
+    const navigate = useNavigate();
     const { userInfo } = useContext(AuthContext);
 
     const [user] = userInfo;
 
-    const {response, isLoading, error, reFetch} = useFetch(`/chatroom/users/${user.id}`);
-
     const [chats, setChats] = useState<Array <ChatInterface>>([]);
+    const [fetchURL, setFetchURL] = useState("");
+
+    const {response, isLoading, error, reFetch} = useFetch(fetchURL);
+
+    useEffect(() => {
+        if(!user) {
+            navigate('/');
+            return
+        }
+        setFetchURL(`/chatroom/users/${user.id}`);
+    }, [])
 
     useEffect(() => {
         if(response) {
@@ -23,8 +33,8 @@ const Dashboard = () => {
 
     return(
         <div className="dashBoardPage">
-            <ConversationList chats={chats} />
-            <Outlet />
+            {user && <ConversationList chats={chats} />}
+            {user && <Outlet />}
         </div>
     )
 }
