+++ src/components/dashboard/list/conversation-list.css
@@ -1,0 +1,12 @@
+.conversationBanner {
+    border: 2px solid blue;
+    width: 100%;
+    height: 5%;
+    display: flex;
+    align-items: center;
+    justify-content: space-around;
+}
+
+.createConversationBtn {
+    margin-top: 1rem;
+}
