+++ src/components/dashboard/room/messages/messages.css
@@ -1,0 +1,7 @@
+.userMessage {
+    color: blue;
+}
+
+.otherMessage {
+    color: red;
+}
