--- src/components/ui-redo/conversationlist/conversation-list.css
@@ -1,8 +1,0 @@
-.conversationListContainer {
-    border: 2px solid blueviolet;
-    width: 30%;
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-}
