--- src/components/ui-redo/dashboard.css
@@ -1,7 +1,0 @@
-.dashBoardPage {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
