--- src/components/ui-redo/room/conversation.css
@@ -1,5 +1,0 @@
-.conversationContainer {
-    width: 70%;
-    border: 2px solid blue;
-    height: 100%;
-}
