+++ src/index.tsx
@@ -4,14 +4,13 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from "./mainapp/App";
 import SignIn from "./components/auth/sign-in";
 import HomePage from "./components/home/home";
-import Chatroom from "./components/chatroom/chatroom";
 import SignUp from "./components/auth/sign-up";
 import CreateChatroom from "./components/create-chatroom/create-chatroom";
 import NotFound from "./components/errors/not-found";
 import ServerError from "./components/errors/server-error";
 import UserInfo from "./components/user-info/user-info-page";
-import Dashboard from "./components/ui-redo/dashboard";
-import Conversation from "./components/ui-redo/room/conversation";
+import Dashboard from "./components/dashboard/dashboard";
+import Conversation from "./components/dashboard/room/conversation";
 
 const root = ReactDOM.createRoot(document.getElementById("root") as Element);
 root.render(
@@ -24,7 +23,6 @@ root.render(
         </Route>
         <Route path="sign-in" element={<SignIn />} />
         <Route path="sign-up" element={<SignUp />} />
-        <Route path="/chat/:chatId" element={<Chatroom />} />
         <Route path="/chat/create" element={<CreateChatroom />} />
         <Route path="/user/:userId" element={<UserInfo />} />
         <Route path="/error" element={<ServerError />} />
